Fixed a crashing bug in Shape::SphereHit() that was introduced recently.
Extended level sphere and ray hit functions to return the object that was hit and also allow an object to be ignored to be specified (this is normally used when hit checking a missile, or whatever, against everything in the level and you don't want it to hit itself).

Made Bullet, Missile and ShipPlayer physics use the above.